+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## v.0.0.20
+
+- Support for finding "multi-package" plugins, where one directory contains
+  multiple packages for different platform implementations.
+
 ## v.0.0.19+3
 
 - Use `package:file` for file I/O.